+++ include/libpkgstate-build/libpkgstate-build.h
@@ -2,4 +2,6 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 #pragma once
 
+#include <libpkgstate-build/export.h>
+
 #include <libpkgstate-build/adapter.h>